00_GCP_Projects/05_GCP_Document_Streaming/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
confluent-kafka==2.6.0
Deprecated==1.2.14
deprecation==2.1.0
fastapi==0.115.4
Flask==3.0.3
functions-framework==3.8.1
google-api-core==2.21.0
google-auth==2.35.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-managedkafka==0.1.4
google-cloud-secret-manager==2.21.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
loguru==0.7.2
MarkupSafe==3.0.2
numpy==2.1.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
Werkzeug==3.0.6
win32-setctime==1.1.0
wrapt==1.16.0
zipp==3.20.2
